requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
apscheduler>=3.10.4
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
certifi>=2024.2.2
playwright>=1.42.0
